Complete refactoring of Sparkfun driver
- Removed all items in implementation that were removed from header
- Adjusted begin() to match behavior of beginI2C() and beginSPI(), which is the version the users were actually expected to use
- Added new abstractions for SPI/I2C communication and the mechanisms to supply them

One note: avr-gcc does not supply the cpp variants of the standard library headers (cstddef vs stddef.h). The cpp headers may include additional details beyond just purely forwarding to the C-style header, so we should default to including that version. For AVR, we use a compiler macro to determine whether we should fall back to the C versions.