Differentiate header includes for avr-gcc
avr-gcc does not supply the cpp variants of the standard library headers (cstddef vs stddef.h). The cpp headers may include additional details beyond just purely forwarding to the C-style header, so we should default to including that version. For AVR, we use a compiler macro to determine whether we should fall back to the C versions.